requirements.txt
Flask~=3.1.0
pandas~=2.2.3
Pillow~=11.0.0
PyMuPDF
python-docx
docx2pdf~=0.1.8
pypdf2~=3.0.1
fpdf~=1.7.2
fitz~=0.0.1.dev2
streamlit